Programmers_부족한금액계산하기

+++ src/year_2022_month_5/Programmers_부족한금액계산하기.java
@@ -1,7 +1,7 @@
 package year_2022_month_5;
 
 public class Programmers_부족한금액계산하기 {
-    public long solution(int price, int money, int count) {
+    public long solution(int price, long money, int count) {
         long answer = -1;
 
         for (int i = 1; i <= count; i++) {
